% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logratio.R
\name{logratio}
\alias{logratio}
\title{Calculate the log-ratio for compositional data}
\usage{
logratio(CC)
}
\arguments{
\item{CC}{A matrix or data.frame whose rows sum to one or one hundred.}
}
\value{
A matrix composed of the log ratios values.
}
\description{
Take a matrix or data.frame whose rows add to one and returns a matrix which is not of compositional data, but its log ratio applied to the rows
}
\examples{
C = matrix(1:20,4)/rowSums(matrix(1:20,4))
logratio(C)
}
